+++ src/main/java/com/example/bookstore/service/LanguageService.java
@@ -7,20 +7,20 @@ import com.example.bookstore.persistance.entity.Language;
 import com.example.bookstore.persistance.repository.LanguageRepository;
 import com.example.bookstore.service.dto.LanguageDTO;
 import com.example.bookstore.service.dto.LanguageRequestDTO;
-import com.example.bookstore.service.mapper.LanguageMapper;
+import com.example.bookstore.service.mapper.Mapper;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class LanguageService {
 
     private final LanguageRepository languageRepository;
-    private final LanguageMapper languageMapper;
+    private final Mapper<Language, LanguageDTO> languageMapper;
 
     public LanguageDTO getById(Long id) {
         return languageMapper.entityToDto(languageRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Language", id)));
@@ -29,16 +29,10 @@ public class LanguageService {
 
     public void delete(Long id) {
         Language language = languageRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Language", id));
-        try {
-            languageRepository.delete(language);
-        }catch (DataIntegrityViolationException e) {
-            if(e.getMostSpecificCause().getMessage().contains("violates foreign key constraint")){
-                List<Long> dependentBookIds = languageRepository.dependentBookIds(id);
-                throw new EntityDeletionException("Language with id: '" + id +
-                        "' could not be deleted successfully. Details: The books with the following ids %s are depended on the specified language.".formatted(dependentBookIds));
-            }
-            throw new EntityDeletionException(id, e.getMessage());
-        }
+        List<Long> dependentBookIds = languageRepository.dependentBookIds(id);
+        if(!dependentBookIds.isEmpty())
+            throw new EntityDeletionException("Language", id, dependentBookIds);
+        languageRepository.delete(language);
     }
 
     public LanguageDTO updateById(Long id, LanguageRequestDTO languageDTO) {
@@ -50,11 +44,16 @@ public class LanguageService {
     public LanguageDTO create(@Valid LanguageRequestDTO languageDTO) {
         Language language = languageRepository.findByLanguage(languageDTO.getName());
         if(language != null){
-            throw new EntityAlreadyExistsException("Language with name '%s' already exists.".formatted(languageDTO.getName()));
+            throw new EntityAlreadyExistsException("Language", "name '%s'".formatted(languageDTO.getName()));
         }
         language = new Language();
         language.setLanguage(languageDTO.getName());
         return languageMapper.entityToDto(languageRepository.save(language));
 
+    }
+
+    public List<LanguageDTO> getAll() {
+        List<Language> languages = languageRepository.findAll();
+        return languages.stream().map(languageMapper::entityToDto).collect(Collectors.toList());
     }
 }
